Add files via upload
This project aims to develop a movie ad campaign
that maximizes profit for an advertising company by determining
which keywords have the greatest impact on a movie’s profitability.
The study uses over 10,000 data points and limited
features to prevent false positives. The research focuses on how
the keywords associated with each movie correlate with profit to
make predictions of which movies are most likely to be profitable.
Four algorithms were used to correlate key words with profit
including a decision tree, random forest, neural network, and
linear regression.